feat: break common features into baseIcon
BREAKING CHANGE: This feature and other related are redefining the scope
of the auro-icon feature set and creates a new auro-alaska CE
for specific Alaska logo support
