+++ docs/api.md
@@ -1,3 +1,26 @@
+# auro-alaska
+
+auro-alaska provides users a way to use the Auro Icons by simply passing in the category and name.
+
+## Properties
+
+| Property         | Attribute        | Type      | Description                                      |
+|------------------|------------------|-----------|--------------------------------------------------|
+| `alaska`         | `alaska`         | `Boolean` | Set value for default alaska airlines logo.      |
+| `alaskaTagline`  | `alaskaTagline`  | `Boolean` | Set value for alaska airlines logo with tagline. |
+| `customColor`    | `customColor`    | `boolean` |                                                  |
+| `hidden`         | `hidden`         | `Boolean` | If present, the component will be hidden both visually and from screen readers |
+| `hiddenAudible`  | `hiddenAudible`  | `Boolean` | If present, the component will be hidden from screen readers, but seen visually |
+| `hiddenVisually` | `hiddenVisually` | `Boolean` | If present, the component will be hidden visually, but still read by screen readers |
+| `onDark`         | `onDark`         | `boolean` |                                                  |
+
+## Slots
+
+| Name | Description                                      |
+|------|--------------------------------------------------|
+|      | Hidden from visibility, used for a11y if icon description is needed |
+
+
 # auro-icon
 
 auro-icon provides users a way to use the Auro Icons by simply passing in the category and name.
