+++ package.json
@@ -94,7 +94,7 @@
   ],
   "scripts": {
     "build": "npm run ciBuild",
-    "apiBuild": "wca analyze 'src/auro-icon.js' --outFiles docs/api.md",
+    "apiBuild": "wca analyze 'src/**/auro-*.js' --outFiles docs/api.md",
     "bundler": "rollup -c",
     "bundle:test": "rollup -c -w",
     "postinstall": "node packageScripts/postinstall.js",
